+++ src/Sulu/Bundle/ContentBundle/Content/Structure/ExcerptStructureExtension.php
@@ -16,13 +16,15 @@ use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
 use Sulu\Component\Content\ContentTypeManagerInterface;
+use Sulu\Component\Content\Export\ContentExportManagerInterface;
 use Sulu\Component\Content\Extension\AbstractExtension;
+use Sulu\Component\Content\Extension\ExportExtensionInterface;
 use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
 
 /**
  * extends structure with seo content.
  */
-class ExcerptStructureExtension extends AbstractExtension
+class ExcerptStructureExtension extends AbstractExtension implements ExportExtensionInterface
 {
     /**
      * name of structure extension.
@@ -56,6 +58,11 @@ class ExcerptStructureExtension extends AbstractExtension
     protected $structureManager;
 
     /**
+     * @var ContentExportManagerInterface
+     */
+    protected $contentExportManager;
+
+    /**
      * @var string
      */
     private $languageNamespace;
@@ -73,11 +80,13 @@ class ExcerptStructureExtension extends AbstractExtension
     public function __construct(
         StructureManagerInterface $structureManager,
         ContentTypeManagerInterface $contentTypeManager,
-        Factory $factory
+        Factory $factory,
+        ContentExportManagerInterface $contentExportManager
     ) {
         $this->contentTypeManager = $contentTypeManager;
         $this->structureManager = $structureManager;
         $this->factory = $factory;
+        $this->contentExportManager = $contentExportManager;
     }
 
     /**
@@ -225,5 +234,33 @@ class ExcerptStructureExtension extends AbstractExtension
         foreach ($this->getExcerptStructure($locale)->getProperties() as $property) {
             $this->properties[] = $property->getName();
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function export($properties, $format = null)
+    {
+        $container = new ExcerptValueContainer($properties);
+
+        $data = [];
+        foreach ($this->getExcerptStructure()->getProperties() as $property) {
+            if ($container->__isset($property->getName())) {
+                $property->setValue($container->__get($property->getName()));
+                $contentType = $this->contentTypeManager->get($property->getContentTypeName());
+                if ($this->contentExportManager->hasExport($property->getContentTypeName(), $format)) {
+                    $options = $this->contentExportManager->getOptions($property->getContentTypeName(), $format);
+
+                    $data[$property->getName()] = [
+                        'name' => $property->getName(),
+                        'value' => $contentType->exportData($property->getValue()),
+                        'type' => $property->getContentTypeName(),
+                        'options' => $options,
+                    ];
+                }
+            }
+        }
+
+        return $data;
     }
 }