+++ src/Sulu/Bundle/ContentBundle/DependencyInjection/SuluContentExtension.php
@@ -111,6 +111,7 @@ class SuluContentExtension extends Extension implements PrependExtensionInterfac
         $loader->load('compat.xml');
         $loader->load('document.xml');
         $loader->load('serializer.xml');
+        $loader->load('export.xml');
     }
 
     private function processPreview(ContainerBuilder $container, $config)