+++ src/Sulu/Bundle/MediaBundle/Content/Types/MediaSelectionContentType.php
@@ -17,12 +17,13 @@ use Sulu\Bundle\MediaBundle\Media\Manager\MediaManagerInterface;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\ComplexContentType;
+use Sulu\Component\Content\ContentTypeExportInterface;
 use Sulu\Component\Util\ArrayableInterface;
 
 /**
  * content type for image selection.
  */
-class MediaSelectionContentType extends ComplexContentType
+class MediaSelectionContentType extends ComplexContentType implements ContentTypeExportInterface
 {
     /**
      * @var MediaManagerInterface
@@ -203,5 +204,22 @@ class MediaSelectionContentType extends ComplexContentType
     public function getViewData(PropertyInterface $property)
     {
         return $property->getValue();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function exportData($propertyValue)
+    {
+        if (is_string($propertyValue)) {
+            return $propertyValue;
+        }
+        if (is_array($propertyValue)) {
+            if (!empty($propertyValue)) {
+                return json_encode($propertyValue);
+            }
+        }
+
+        return '';
     }
 }